backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
requests==2.31.0
Pillow==10.4.0
